vending_machine/requirements.txt
streamlit
kafka-python-ng
dlt[postgres]
psycopg2-binary
duckdb
pandas